+++ js/data.js
@@ -1,0 +1,42 @@
+// import {comments, posts} from './util.js';
+
+//Oписание фотографии.
+const descriptions = [
+  'Лицо человека, вышедшего из матрицы',
+  'В этот день родили меня на свет...',
+  'Если не поставишь лайк, я тебя найду!',
+  'Аниме, пацаны?',
+  'Тот момент, когда не ожидаешь',
+  'Просто вчера на все 100',
+  'Подумал, что не хватит, купил ещё',
+  'Мой лучший друг идиот',
+  'Не поделили еду с котом',
+];
+
+//Сообщения от комментатора
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const names = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+const minLikesAmount = 15;
+const maxLikesAmount = 200;
+const minAvatarNumber = 1;
+const maxAvatarNumber = 6;
+
+export {descriptions, messages, names, minLikesAmount, maxLikesAmount, minAvatarNumber, maxAvatarNumber};
